Rename README.md to LICENSE
Copyright (c) <year> <copyright holders> "Anti 996" License Version 1.0 (Draft) Permission is hereby granted to any individual or legal entityobtaining a copy of this licensed work (including the source code,documentation and/or related items, hereinafter collectively referredto as the "licensed work"), free of charge, to deal with the licensedwork for any purpose, including without limitation, the rights to use,reproduce, modify, prepare derivative works of, distribute, publishand sublicense the licensed work, subject to the following conditions: 1. The individual or the legal entity must conspicuously display,without modification, this License and the notice on each redistributedor derivative copy of the Licensed Work. 2. The individual or the legal entity must strictly comply with allapplicable laws, regulations, rules and standards of the jurisdictionrelating to labor and employment where the individual is physicallylocated or where the individual was born or naturalized; or where thelegal entity is registered or is operating (whichever is stricter). Incase that the jurisdiction has no such laws, regulations, rules andstandards or its laws, regulations, rules and standards areunenforceable, the individual or the legal entity are required tocomply with Core International Labor Standards. 3. The individual or the legal entity shall not induce, suggest or forceits employee(s), whether full-time or part-time, or its independentcontractor(s), in any methods, to agree in oral or written form, todirectly or indirectly restrict, weaken or relinquish his or herrights or remedies under such laws, regulations, rules and standardsrelating to labor and employment as mentioned above, no matter whethersuch written or oral agreements are enforceable under the laws of thesaid jurisdiction, nor shall such individual or the legal entitylimit, in any methods, the rights of its employee(s) or independentcontractor(s) from reporting or complaining to the copyright holder orrelevant authorities monitoring the compliance of the license aboutits violation(s) of the said license. THE LICENSED WORK IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OFMERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE COPYRIGHT HOLDER BE LIABLE FOR ANY CLAIM,DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OROTHERWISE, ARISING FROM, OUT OF OR IN ANY WAY CONNECTION WITH THELICENSED WORK OR THE USE OR OTHER DEALINGS